+++ ewp-api-agent/src/main/java/kr/co/ewp/api/controller/EnergyController.java
@@ -981,7 +981,7 @@ public class EnergyController {
         					  pvGen.setDeviceId(_deviceId);
         					  pvGen.setSiteId(_siteId);
         					  pvGen.setStdDate(item.getTimestamp());
-        					  pvGen.setGenVal(item.getUsage().intValue() / 1000000);
+        					  pvGen.setGenVal((float) (item.getUsage().intValue() / 1000000));
         					  pvGen.setTemp(0);
         					  
         					  pvGentList.add(pvGen);
